requirements.txt
Django==2.0.5
psycopg2-binary
requests_oauthlib
gunicorn
whitenoise
social-auth-app-django
dj-database-url
django-bootstrap4
user-agents
pyyaml
djangorestframework
django-filter

